Modificación del Reporte general en PDF.

+++ src/main/java/umg/principal/DaseDatos/Dao/ProductoDAO.java
@@ -44,7 +44,7 @@ public class ProductoDAO {
              Statement stmt = conn.createStatement();
              ResultSet rs = stmt.executeQuery(sql)) {
             while (rs.next()) {
-                productos.add(new Producto(rs.getInt("id_producto"), rs.getString("descripcion"), rs.getString("origen"), rs.getInt("precio"), rs.getInt("existencia")));
+                productos.add(new Producto(rs.getInt("id_producto"), rs.getString("descripcion"), rs.getString("origen"), rs.getDouble("precio"), rs.getInt("existencia")));
             }
         }
         return productos;
@@ -58,7 +58,7 @@ public class ProductoDAO {
              Statement stmt = conn.createStatement();
              ResultSet rs = stmt.executeQuery(sql)) {
             while (rs.next()) {
-                productos.add(new Producto(rs.getInt("id_producto"), rs.getString("descripcion"), rs.getString("origen"), rs.getInt("precio"), rs.getInt("existencia")));
+                productos.add(new Producto(rs.getInt("id_producto"), rs.getString("descripcion"), rs.getString("origen"), rs.getDouble("precio"), rs.getInt("existencia")));
             }
         }
         return productos;
